+++ common_source_files/clock_calibration.h
@@ -1,7 +1,7 @@
 #ifndef SRC_CLOCK_CALIBRATION_H_
 #define SRC_CLOCK_CALIBRATION_H_
 
-int32_t time_sync(uint64_t time_from_master);
+int32_t time_sync(uint32_t time_from_master);
 uint16_t get_clock_calibration_value(void);
